requirements.txt
streamlit
faiss-cpu
numpy
google-generativeai
sentence-transformers
pypdf
torch